+++ src/pretalx/event/models/event.py
@@ -182,6 +182,8 @@ class Event(LogMixin, models.Model):
         user_submissions = '{user}/submissions'
         schedule = '{base}/schedule'
         sneakpeek = '{base}/sneak'
+        talks = '{base}/talk/'
+        speakers = '{base}/speaker/'
         changelog = '{schedule}/changelog'
         frab_xml = '{schedule}.xml'
         frab_json = '{schedule}.json'