+++ src/tests/functional/agenda/test_agenda_schedule.py
@@ -1,10 +1,7 @@
-import datetime
 from urllib.parse import quote
 
 import pytest
-import pytz
 from django.urls import reverse
-from django.utils import formats
 
 
 @pytest.mark.django_db
@@ -18,160 +15,18 @@ def test_can_see_schedule(client, user, event, slot):
 
 
 @pytest.mark.django_db
-def test_can_see_talk(client, event, slot):
-    response = client.get(slot.submission.urls.public, follow=True)
-    assert event.schedules.count() == 2
+def test_speaker_list(client, event, speaker, slot):
+    response = client.get(event.urls.speakers, follow=True)
     assert response.status_code == 200
-    content = response.content.decode()
-    assert content.count(slot.submission.title) >= 2  # meta+h1
-    assert slot.submission.abstract in content
-    assert slot.submission.description in content
-    assert formats.date_format(slot.start.astimezone(pytz.timezone(event.timezone)), 'Y-m-d, H:i') in content
-    assert formats.date_format(slot.end.astimezone(pytz.timezone(event.timezone)), 'H:i') in content
-    assert str(slot.room.name) in content
-    assert 'fa-edit' not in content  # edit btn
-    assert 'fa-video' not in content  # do not record
-
-
-@pytest.mark.django_db
-def test_cannot_see_new_talk(client, event, unreleased_slot):
-    slot = unreleased_slot
-    response = client.get(slot.submission.urls.public, follow=True)
-    assert event.schedules.count() == 1
-    assert response.status_code == 404
-
-
-@pytest.mark.django_db
-def test_orga_can_see_new_talk(orga_client, event, unreleased_slot):
-    slot = unreleased_slot
-    response = orga_client.get(slot.submission.urls.public, follow=True)
-    assert event.schedules.count() == 1
-    assert response.status_code == 200
-    content = response.content.decode()
-    assert content.count(slot.submission.title) >= 2  # meta+h1
-    assert slot.submission.abstract in content
-    assert slot.submission.description in content
-    assert formats.date_format(slot.start.astimezone(pytz.timezone(event.timezone)), 'Y-m-d, H:i') in content
-    assert formats.date_format(slot.end.astimezone(pytz.timezone(event.timezone)), 'H:i') in content
-    assert str(slot.room.name) in content
-    assert 'fa-edit' not in content  # edit btn
-    assert 'fa-video' not in content  # do not record
-
-
-@pytest.mark.django_db
-def test_can_see_talk_edit_btn(orga_client, orga_user, event, slot):
-    slot.submission.speakers.add(orga_user)
-    response = orga_client.get(slot.submission.urls.public, follow=True)
-    assert response.status_code == 200
-    content = response.content.decode()
-    assert 'fa-edit' in content  # edit btn
-    assert 'fa-video' not in content
-    assert 'fa-comments' not in content
-
-
-@pytest.mark.django_db
-def test_can_see_talk_do_not_record(client, event, slot):
-    slot.submission.do_not_record = True
-    slot.submission.save()
-    response = client.get(slot.submission.urls.public, follow=True)
-    assert response.status_code == 200
-    content = response.content.decode()
-    assert 'fa-edit' not in content  # edit btn
-    assert 'fa-video' in content
-    assert 'fa-comments' not in content
-
-
-@pytest.mark.django_db
-def test_can_see_talk_does_accept_feedback(client, event, slot):
-    slot.start = (datetime.datetime.now() - datetime.timedelta(days=1))
-    slot.end = slot.start + datetime.timedelta(hours=1)
-    slot.save()
-    response = client.get(slot.submission.urls.public, follow=True)
-    assert response.status_code == 200
-    content = response.content.decode()
-    assert 'fa-edit' not in content  # edit btn
-    assert 'fa-comments' in content
-    assert 'fa-video' not in content
-
-
-@pytest.mark.django_db
-def test_cannot_see_nonpublic_talk(client, event, slot):
-    event.is_public = False
-    event.save()
-    response = client.get(slot.submission.urls.public, follow=True)
-    assert response.status_code == 404
-
-
-@pytest.mark.django_db
-def test_cannot_see_other_events_talk(client, event, slot, other_event):
-    response = client.get(slot.submission.urls.public.replace(event.slug, other_event.slug), follow=True)
-    assert response.status_code == 404
-
-
-@pytest.mark.django_db
-def test_event_talk_visiblity_submitted(client, event, submission):
-    response = client.get(submission.urls.public, follow=True)
-    assert response.status_code == 404
-
-
-@pytest.mark.django_db
-def test_event_talk_visiblity_accepted(client, event, slot, accepted_submission):
-    response = client.get(accepted_submission.urls.public, follow=True)
-    assert response.status_code == 404
-
-
-@pytest.mark.django_db
-def test_event_talk_visiblity_confirmed(client, event, slot, confirmed_submission):
-    response = client.get(confirmed_submission.urls.public, follow=True)
-    assert response.status_code == 200
-
-
-@pytest.mark.django_db
-def test_event_talk_visiblity_canceled(client, event, slot, canceled_submission):
-    response = client.get(canceled_submission.urls.public, follow=True)
-    assert response.status_code == 404
-
-
-@pytest.mark.django_db
-def test_event_talk_visiblity_withdrawn(client, event, slot, withdrawn_submission):
-    response = client.get(withdrawn_submission.urls.public, follow=True)
-    assert response.status_code == 404
-
-
-@pytest.mark.django_db
-def test_talk_speaker_other_talks(client, event, speaker, slot, other_slot, other_submission):
-    other_submission.speakers.add(speaker)
-    response = client.get(other_submission.urls.public, follow=True)
-
-    assert response.context['speakers']
-    assert len(response.context['speakers']) == 2, response.context['speakers']
-    speaker_response = [s for s in response.context['speakers'] if s.name == speaker.name][0]
-    other_response = [s for s in response.context['speakers'] if s.name != speaker.name][0]
-    assert len(speaker_response.other_talks) == 1
-    assert len(other_response.other_talks) == 0
-    assert speaker_response.other_talks[0].submission.title == speaker.submissions.first().title
-
-
-@pytest.mark.django_db
-def test_talk_speaker_other_talks_only_if_visible(client, event, speaker, slot, other_slot, other_submission):
-    other_submission.speakers.add(speaker)
-    response = client.get(other_submission.urls.public, follow=True)
-    slot.submission.accept(force=True)
-    slot.is_visible = False
-    slot.save()
-    slot.submission.save()
-
-    assert response.context['speakers']
-    assert len(response.context['speakers']) == 2, response.context['speakers']
-    speaker_response = [s for s in response.context['speakers'] if s.name == speaker.name][0]
-    other_response = [s for s in response.context['speakers'] if s.name != speaker.name][0]
-    assert len(speaker_response.other_talks) == 0
-    assert len(other_response.other_talks) == 0
+    assert speaker.name in response.content.decode()
 
 
 @pytest.mark.django_db
 def test_speaker_page(client, event, speaker, slot):
-    response = client.get(reverse('agenda:speaker', kwargs={'code': speaker.code, 'event': event.slug}), follow=True)
+    response = client.get(
+        reverse('agenda:speaker', kwargs={'code': speaker.code, 'event': event.slug}),
+        follow=True,
+    )
     assert response.status_code == 200
     assert speaker.profiles.get(event=event).biography in response.content.decode()
 
@@ -196,11 +51,7 @@ def test_versioned_schedule_page(client, event, speaker, slot, schedule):
     assert slot.submission.title in response.content.decode()
 
     version = quote(schedule.version)
-    redirected_response = client.get(f'/{event.slug}/schedule?version={version}', follow=True)
+    redirected_response = client.get(
+        f'/{event.slug}/schedule?version={version}', follow=True
+    )
     assert redirected_response._request.path == response._request.path
-
-
-@pytest.mark.django_db
-def test_talk_review_page(client, event, submission, other_submission):
-    response = client.get(submission.urls.review, follow=True)
-    assert response.status_code == 200